+++ student.hpp
@@ -59,7 +59,7 @@ class InternationalStudent : public Student{
 //child class of the Student class
     public:
         InternationalStudent(string FName, string LName, float grade, int RScore, int id, string Con, int TOEFL);
-        InternationalStudent(string FName, string LName, float id, string Con, int TOEFL);
+        InternationalStudent(string FName, string LName, int id, string Con, int TOEFL);
         InternationalStudent();
         void set_Country(string Con);
         void set_TOEFLScore(int TOEFL);
